requirements.txt
fastapi
uvicorn
ollama
chromadb
pypdf
python-dotenv
